Tree view improvements
