+++ app/assets/src/styles/visualization.scss
@@ -262,11 +262,7 @@
     border-radius: 5px;
     box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.2);
     pointer-events: none;
-    font-size: 8pt;
-
-    .name {
-      font-size: 9pt;
-    }
+    border: 1px solid #ccc;
   }
 }
